Updating example for multiple line chart.
