+++ packages/charting/src/components/LineChart/LineChartPage.tsx
@@ -4,9 +4,11 @@ import { ComponentPage, ExampleCard, IComponentDemoPageProps, PropertiesTableSet
 
 import { LineChartBasicExample } from './examples/LineChart.Basic.Example';
 import { LineChartStyledExample } from './examples/LineChart.Styled.Example';
+import { LineChartMultipleExample } from './examples/LineChart.Multiple.Example';
 
 const LineChartBasicExampleCode = require('!raw-loader!@uifabric/charting/src/components/LineChart/examples/LineChart.Basic.Example.tsx') as string;
 const LineChartStyledExampleCode = require('!raw-loader!@uifabric/charting/src/components/LineChart/examples/LineChart.Styled.Example.tsx') as string;
+const MultipleLineChartExampleCode = require('!raw-loader!@uifabric/charting/src/components/LineChart/examples/LineChart.Multiple.Example.tsx') as string;
 
 export class LineChartPage extends React.Component<IComponentDemoPageProps, {}> {
   public render(): JSX.Element {
@@ -21,6 +23,9 @@ export class LineChartPage extends React.Component<IComponentDemoPageProps, {}> 
             </ExampleCard>
             <ExampleCard title="LineChart styled" code={LineChartStyledExampleCode}>
               <LineChartStyledExample />
+            </ExampleCard>
+            <ExampleCard title="Multiple Line chart" code={MultipleLineChartExampleCode}>
+              <LineChartMultipleExample />
             </ExampleCard>
           </div>
         }
