+++ packages/charting/src/components/LineChart/examples/LineChart.Basic.Example.tsx
@@ -25,37 +25,9 @@ export class LineChartBasicExample extends React.Component<ILineChartProps, {}> 
         { x: 40, y: 29 },
         { x: 45, y: 43 },
         { x: 50, y: 45 }
-      ],
-      [
-        { x: 0, y: 18 },
-        { x: 5, y: 20 },
-        { x: 10, y: 40 },
-        { x: 15, y: 30 },
-        { x: 20, y: 18 },
-        { x: 22, y: 20 },
-        { x: 35, y: 40 },
-        { x: 40, y: 30 },
-        { x: 42, y: 18 },
-        { x: 43, y: 20 },
-        { x: 45, y: 40 },
-        { x: 50, y: 30 }
-      ],
-      [
-        { x: 0, y: 20 },
-        { x: 5, y: 15 },
-        { x: 10, y: 30 },
-        { x: 15, y: 35 },
-        { x: 20, y: 30 },
-        { x: 22, y: 15 },
-        { x: 35, y: 30 },
-        { x: 40, y: 27 },
-        { x: 42, y: 29 },
-        { x: 43, y: 35 },
-        { x: 45, y: 40 },
-        { x: 50, y: 42 }
       ]
     ];
-    const colors = ['#FF5733', '#176213', '#193BBD'];
+    const colors = ['#FF5733'];
     return <LineChart data={points} chartLabel={'Basic Chart with Numeric Axes'} colors={colors} />;
   }
 }
