+++ packages/charting/src/components/LineChart/examples/LineChart.Multiple.Example.tsx
@@ -3,7 +3,7 @@ import { LineChart, ILineChartProps } from '@uifabric/charting/lib/LineChart';
 import { customizable } from 'office-ui-fabric-react/lib/Utilities';
 import { IDataPoint } from '@uifabric/charting/lib/types/IDataPoint';
 @customizable('HorizontalBarChartBasicExample', ['theme', 'styles'])
-export class LineChartStyledExample extends React.Component<ILineChartProps, {}> {
+export class LineChartMultipleExample extends React.Component<ILineChartProps, {}> {
   constructor(props: ILineChartProps) {
     super(props);
   }
@@ -18,19 +18,47 @@ export class LineChartStyledExample extends React.Component<ILineChartProps, {}>
 
     const points: IDataPoint[][] = [
       [
-        { x: 'One', y: 5 },
-        { x: 'Two', y: 10 },
-        { x: 'Three', y: 5 },
-        { x: 'Four', y: 20 },
-        { x: 'Five', y: 18 },
-        { x: 'Six', y: 5 },
-        { x: 'Seven', y: 10 },
-        { x: 'Eight', y: 5 },
-        { x: 'Nine', y: 20 },
-        { x: 'Ten', y: 18 }
+        { x: 0, y: 10 },
+        { x: 5, y: 18 },
+        { x: 10, y: 24 },
+        { x: 15, y: 25 },
+        { x: 20, y: 15 },
+        { x: 25, y: 30 },
+        { x: 30, y: 18 },
+        { x: 35, y: 32 },
+        { x: 40, y: 29 },
+        { x: 45, y: 43 },
+        { x: 50, y: 45 }
+      ],
+      [
+        { x: 0, y: 18 },
+        { x: 5, y: 20 },
+        { x: 10, y: 40 },
+        { x: 15, y: 30 },
+        { x: 20, y: 18 },
+        { x: 22, y: 20 },
+        { x: 35, y: 40 },
+        { x: 40, y: 30 },
+        { x: 42, y: 18 },
+        { x: 43, y: 20 },
+        { x: 45, y: 40 },
+        { x: 50, y: 30 }
+      ],
+      [
+        { x: 0, y: 20 },
+        { x: 5, y: 15 },
+        { x: 10, y: 30 },
+        { x: 15, y: 35 },
+        { x: 20, y: 30 },
+        { x: 22, y: 15 },
+        { x: 35, y: 30 },
+        { x: 40, y: 27 },
+        { x: 42, y: 29 },
+        { x: 43, y: 35 },
+        { x: 45, y: 40 },
+        { x: 50, y: 42 }
       ]
     ];
-
     const axisStyle = {
       stroke: palette.purpleDark
     };
@@ -61,7 +89,7 @@ export class LineChartStyledExample extends React.Component<ILineChartProps, {}>
       };
     };
 
-    const colors = ['#FF5733'];
+    const colors = ['#FF5733', '#176213', '#193BBD'];
 
     return (
       <LineChart
